Pausing system implemented
Pausing via the LevelManager is now implemented. Will
need to add calls to the pause and unpause methods to
keep up with more entities, but there should not be too
many types to manage. Currently no pause menu, but game
logic stops and starts via state machines instead of time
scale, which frees up time scale for special effects in
the future!