Camera controller improved
Camera has been improved for moving in a 3D space, now
follows the player in a logical way when moving around.
Known issues are a literal edge case in which leaving
the move range while falling causes the move range to fall with the
player and the camera jerking around as the player rotates
around the move range.