Improved camera vertical control
Camera now has vertical angle limits, but there
is a known issue in which the limits are violated.
The only implementation explored was jerky, so this
remains a work in progress.

Also added variable velocity jump, which changes how
high the player jumps based on how long the jump
button is held.