Camera now functional, but not perfect
The camera can now do everything it needs to
aside from locking on and special cases, but
some transitions (particularly long falls
in follow mode and the peak of a jump) are
somewhat abrupt. This is lower priority for now
as the functionality is all there, though the
transitions could be more smooth.

The Sphere Prefab was also removed to fix a build
warning.