More work done on the combat controller
The combat controller now has more of its structure coming together. The
master controller now updates it when attacking is active, which should
mean it is essentially "short circuited" when not active. This is
untested, but the whole system doesn't seem too terribly complicated so
it should not be too buggy by the time animations are ready.