Combat controller and weapons now function on a basic level
Through a variety of debugging methods, a simple two-hit "combo" is now
in the game. Without any animations, there isn't much visual feedback so
the next step would be working on animations. The backend looks a tad
messy thanks to the currently-bespoke data structure used to structure
combos, but setting everything up should be relatively simple, though I
think I will implement a Serialized version of the Attack class to make
adding attacks a drag-and-drop affair. The bits and pieces are there,
meaning that the "backend" functionality of the player character is at a
functional level.