Continued testing and set up of combat controller
Replaced the inelegant booleans with a state machine for the weapon and
attack structure. The combat controller (thankfully) plays nicely with
multiple attacks in a single string. It would be worth experimenting
with branching strings sooner rather than later, but until charging and
delayed combos are desired, the next step for combat controls is
animation