Added a new model and fixed a notorious camera glitch
The camera has an outdated state that caused a glitch when run outside
the editor. For some reason, this problem did not exist in the editor.
The old state and associated code will be pruned later, but currently
the camera works correctly from the start in the built game. Also
included was a basic sword weapon for use in animation/combat in the
future.