Working anmiation and fixed character model, starting event system.
The unity event system should make so much of the architecture of the
game simpler. It already seems a bit tangled so leveraging a way to
signal an event and have listeners react accordingly should make this
whole deal a lot easier. However, the immediate change is the addition
of a character model that moves! It animates when walking, but the
animation is very much programmer art. New animations will be added,
which is the next step to working combat.